requirements.txt
numpy>=1.19.0
matplotlib>=3.3.0
Pillow>=8.0.0
opencv-python>=4.5.0
scikit-image>=0.18.0
trimesh>=3.9.0